backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
qrcode==7.4.2
django-extensions==3.2.3
pyzbar==0.1.9
